M1: quick wins
Audit-driven tech-debt cleanup (Linear SMA-328..SMA-333), one milestone
PR per the Optimisations project. Low-risk, high-confidence fixes.

- SMA-328: extract the duplicated wait_bounded subprocess reaper (child
  wait + poll + timeout-kill) from render.rs and update/mod.rs into one
  shared src/proc.rs helper. Pure dedup, no behavior change (4/4-model
  finding, security-adjacent).
- SMA-329: correct the stale prebuilt-gate comment in release.yml — the
  install step selects the prebuilt by declared version match, not commit
  exactness; COMMIT is informational only.
- SMA-330: ARCHITECTURE.md now notes the one on-disk exception to
  'ephemeral state only' — the advisory update-check.json cache.
- SMA-331: docs/install.md lists git as a runtime requirement (the
  git-aware tree + diff views shell out to git; degraded without it).
- SMA-332: swept src/** comments clean of build-process refs (T-N,
  SMA-N, review-gate R-N) and corrected the stale 'search keystrokes are
  no-ops' comment in controller.rs (search is fully implemented).
- SMA-333: synced lagging doc lines — herdr-plugin.toml pane comment,
  CHANGELOG missing release-link refs, docs/renderers.md bundled palette
  + cargo-fallback correction, docs/install.md NO_UPDATE_CHECK wording
  (any value disables, not just =1). Local-only docs (CONTEXT.md, ADR
  0002, AGENTS.md) fixed on disk.

DoD: CHANGELOG [Unreleased], ARCHITECTURE.md module table (+proc).

Green bar: cargo build --release, cargo test, cargo clippy --all-targets
-- -D warnings, cargo fmt --check all clean.
