+++ docs/renderers.md
@@ -10,7 +10,9 @@ dependencies — not Cargo dependencies — and each is **optional**:
 | Syntax-highlighted content | [`bat`](https://github.com/sharkdp/bat) | `brew install bat` / package manager |
 
 Or install all three at once with the bundled helper (best-effort; detects brew/apt/dnf/pacman
-and falls back to `cargo`), run from the plugin dir (`herdr plugin list` shows its path):
+and falls back to `cargo install` for `delta` and `bat` — `glow` is written in Go, so the helper
+prints its manual install link instead of attempting a cargo install), run from the plugin dir
+(`herdr plugin list` shows its path):
 
 ```bash
 ./scripts/install-renderers.sh
@@ -25,3 +27,12 @@ the best experience but not required to use the viewer.
 Untrusted file content is always fed to a renderer on **stdin** (never as a command argument),
 and the renderer's output is re-sanitized before display, so a hostile file name or file
 content cannot inject a command or drive the terminal.
+
+### Bundled markdown palette
+
+The viewer ships a small bundled markdown style palette (`assets/markdown-style.json`) that
+`glow` is pointed at when it is present, so rendered markdown uses a consistent set of named
+ANSI colors (headings, code blocks, links, etc.) rather than glow's built-in `dark` style.
+When the palette file is absent, glow falls back to its built-in `dark` style — markdown still
+renders, just with glow's default colors. The palette is a trusted glow argument (located only
+inside the plugin's own dirs), never derived from untrusted input.
