+++ src/context.rs
@@ -1,6 +1,6 @@
 //! The normalized launch context the Host Adapter hands to the Root Resolver.
 //!
-//! Produced at the herdr boundary (T-17) from injected env/JSON; consumed by
+//! Produced at the herdr boundary from injected env/JSON; consumed by
 //! [`crate::root::resolve`]. Malformed host input degrades to a minimal `{ cwd }`.
 
 use std::path::PathBuf;
